Update matcher's internal discriminant during widget update (#13)
Currently the matcher's discriminant is only updated during lifecycle
calls. However, this only works reliably when switching for the first
time since only then the LifeCycle::WidgetAdded is called. When
switching to a certain variant a second time the lifecycle method is
not necessarily called and this means the matcher and the data are
then out of sync.

The most sensible place to handle this change from one variant
to another is in the update function. If the data changes also the
used subwidget should change accordingly.